+++ cmd/ttn-lw-cli/commands/clients_access.go
@@ -78,8 +78,9 @@ var (
 		},
 	}
 	clientCollaboratorsSet = &cobra.Command{
-		Use:   "set",
-		Short: "Set a client collaborator",
+		Use:     "set",
+		Aliases: []string{"update"},
+		Short:   "Set a client collaborator",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cliID := getClientID(cmd.Flags(), nil)
 			if cliID == nil {
@@ -114,7 +115,7 @@ var (
 	}
 	clientCollaboratorsDelete = &cobra.Command{
 		Use:     "delete",
-		Aliases: []string{"remove"},
+		Aliases: []string{"del", "remove", "rm"},
 		Short:   "Delete a client collaborator",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cliID := getClientID(cmd.Flags(), nil)