+++ cmd/ttn-lw-cli/commands/end_devices.go
@@ -179,6 +179,7 @@ var (
 	}
 	endDevicesListFrequencyPlans = &cobra.Command{
 		Use:               "list-frequency-plans",
+		Aliases:           []string{"get-frequency-plans", "frequency-plans", "fps"},
 		Short:             "List available frequency plans for end devices",
 		PersistentPreRunE: preRun(),
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -554,9 +555,9 @@ var (
 		}),
 	}
 	endDevicesUpdateCommand = &cobra.Command{
-		Use:     "update [application-id] [device-id]",
-		Aliases: []string{"set"},
-		Short:   "Update an end device",
+		Use:     "set [application-id] [device-id]",
+		Aliases: []string{"update"},
+		Short:   "Set properties of an end device",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			forwardDeprecatedDeviceFlags(cmd.Flags())
 
@@ -769,8 +770,9 @@ var (
 		},
 	}
 	endDevicesDeleteCommand = &cobra.Command{
-		Use:   "delete [application-id] [device-id]",
-		Short: "Delete an end device",
+		Use:     "delete [application-id] [device-id]",
+		Aliases: []string{"del", "remove", "rm"},
+		Short:   "Delete an end device",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			devID, err := getEndDeviceID(cmd.Flags(), args, true)
 			if err != nil {
@@ -923,7 +925,7 @@ values will be stored in the Join Server.`,
 	}
 	endDevicesListQRCodeFormatsCommand = &cobra.Command{
 		Use:     "list-qr-formats",
-		Aliases: []string{"ls-qr-formats", "listqrformats", "lsqrformats", "lsqrfmts", "lsqrfmt"},
+		Aliases: []string{"ls-qr-formats", "listqrformats", "lsqrformats", "lsqrfmts", "lsqrfmt", "qr-formats"},
 		Short:   "List QR code formats (EXPERIMENTAL)",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			qrg, err := api.Dial(ctx, config.QRCodeGeneratorGRPCAddress)